Compiler suggested fixes

+++ PDP8/pdp8_ct.c
@@ -717,7 +717,7 @@ static const uint16 boot_rom[] = {
 
 t_stat ct_boot (int32 unitno, DEVICE *dptr)
 {
-int32 i;
+size_t i;
 extern int32 saved_PC;
 extern uint16 M[];
